Deleted obsolete Move class, the related blackboard notifications.
Blackboard notifications are not longer used to publish mobility state.
changes. Instead INET mobility models provide accurate positions at any
simulation time. It is no longer needed to store a linear approximation
of the hosts's movement. This was bad for two reasons: it was just an
approximation and the user had to configure an ad-hoc time interval for it.

Also replaced the old blackboard publish/subscribe for position update
with the OMNeT standard signaling mechanism. This is still used by the
ConnectionManager class.