fix: workaround for a crash when the the wakeup timer was deleted.
The timer should never be deleted as it is a self message. This commit should
be verified by someone, who knows a bit more about LMAC as there are definitely
unhandled cases when a wakeup timer arrives in a state that does not handle it.
Till now the default handling was to delete the message. We are just skipping it
now.